Use generic function to retrieve machines
